Fix partialy Dev-542 of newer binary upload format

+++ src/test/java/com/kloudless/KloudlessTest.java
@@ -235,7 +235,7 @@ public class KloudlessTest extends KloudlessBaseTest {
         e.printStackTrace();
       }
       KloudlessResponse response = File.contents(file.id, account,null);
-      int actualSize = response.getResponseStream().size();
+      long actualSize = response.getResponseStream().size();
       assertThat(actualSize).isEqualTo(expectedSize);
     }
     //TODO: text file
